+++ xarray/plot/utils.py
@@ -3,6 +3,9 @@ from __future__ import absolute_import, division, print_function
 import warnings
 
 import numpy as np
+import pandas as pd
+import pkg_resources
+import textwrap
 
 from ..core.pycompat import basestring
 from ..core.utils import is_scalar
@@ -334,3 +337,24 @@ def get_axis(figsize, size, aspect, ax):
         ax = plt.gca()
 
     return ax
+
+
+def label_from_attrs(da):
+    ''' Makes informative labels if variable metadata (attrs) follows
+        CF conventions. '''
+
+    if da.attrs.get('long_name'):
+        name = da.attrs['long_name']
+    elif da.attrs.get('standard_name'):
+        name = da.attrs['standard_name']
+    elif da.name is not None:
+        name = da.name
+    else:
+        name = ''
+
+    if da.attrs.get('units'):
+        units = ' [{}]'.format(da.attrs['units'])
+    else:
+        units = ''
+
+    return '\n'.join(textwrap.wrap(name + units, 30))